requirements.txt
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.2.0
django-extensions==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
entrypoints==0.3
idna==2.8
inflection==0.3.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pdfminer.six==20181108
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycryptodome==3.8.2
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
pyrsistent==0.15.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.2
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.0
